requirements.txt
paho-mqtt
jinja2
scipy
python-dotenv
fastapi
discord.py
